pkg: controllers: move MCP code in its own package
Extract MCP utilities from the controllers to its own package,
so the test code doesn't depend on the controllers anymore.
